+++ CHANGELOG.md
@@ -18,6 +18,8 @@
 * Double decryption when reconstructing invalid files
 * Parsing encrypted indirect objects that should be streams but aren't
 * Serializing of large floats that used scientific notation
+* [HexaPDF::Content::SmartTextExtractor::layout_text_runs] to work in case of
+  degenerate input where the majority of text runs have zero height
 
 
 ## 1.9.1 - 2026-06-09